% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allterms.R
\name{all_terms_level_h2}
\alias{all_terms_level_h2}
\title{Building tree}
\usage{
all_terms_level_h2(xx.ch, terms)
}
\arguments{
\item{xx.ch}{The list Go child terms}

\item{terms}{The Go terms whoes children need to be determine}
}
\value{
returns GO terms.
}
\description{
Building tree
}
\keyword{internal}
