% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeDistBP.R
\name{degreeDistBP}
\alias{degreeDistBP}
\title{Degree distribution of the GO biological process (BP) terms on a GO-level}
\usage{
degreeDistBP(level)
}
\arguments{
\item{level}{A numeric value for the GO-level}
}
\value{
A plot showing the degree distribution
}
\description{
For a directed graph, the in-degree nodes are the nodes which have edges coming into them
             and the out-degree nodes are those which have edges going out of them. The degreeDistBP
             function shows the distribution of these degrees over a particular GO-level. A bar plot is obtain
             which shows how many nodes in the GO-level have a certain degree k.
}
\examples{
# Degree distribution of GO-terms on level 3
degreeDistBP(level = 3)

# Degree distribution of GO-terms on level 9
degreeDistBP(level = 9)

}
