% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawGraph.R
\name{drawGraph}
\alias{drawGraph}
\title{Visualize GO DAG}
\usage{
drawGraph(Elegans, nodeinfo, title, species, ont, org)
}
\arguments{
\item{Elegans}{The igraph object}

\item{nodeinfo}{The graph structure}

\item{title}{The title of the plot}

\item{species}{The organism of interest}

\item{ont}{The ontology}

\item{org}{Organism}
}
\value{
A graph of the simplified GO tree
}
\description{
Visualize GO DAG
}
\keyword{internal}
