% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawGraph4vis.R
\name{drawGraph4Vis}
\alias{drawGraph4Vis}
\title{Visualize GO DAG}
\usage{
drawGraph4Vis(Elegans, nodeinfo, title, species, ont, val, org)
}
\arguments{
\item{Elegans}{The igraph object}

\item{nodeinfo}{The graph structure}

\item{title}{The title of the plot}

\item{species}{The organism of interest}

\item{ont}{The ontology}

\item{val}{The important GO categories}

\item{org}{The organism}
}
\value{
A graph of the simplified GO tree
}
\description{
Visualize GO DAG
}
\keyword{internal}
