% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allterms.R
\name{getGOcategory}
\alias{getGOcategory}
\title{GO-term category getter}
\usage{
getGOcategory(goterm)
}
\arguments{
\item{goterm}{An object with a category}
}
\value{
A data.frame object containing the terms, category and ontology.
}
\description{
Get the category of a GO-term. It checks if a GO-term is a jump node (RN), regular node (RN) or
leaf node (LN).
}
\examples{
goterm <- c("GO:0106003","GO:0032991","GO:1990429","GO:0002133","GO:0089713",
"GO:1990666","GO:0036125")


getGOcategory(goterm)

}
