% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allterms.R
\name{getNodeTerms}
\alias{getNodeTerms}
\title{Retrieve GO terms in each category}
\usage{
getNodeTerms(term, v)
}
\arguments{
\item{term}{The skeleton of the reduced DAG}

\item{v}{The actual GO terms}
}
\value{
Returns a list of GO terms in each node category
}
\description{
Retrieve GO terms in each category
}
\keyword{internal}
