% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeListBp.R
\name{GO2DecBP}
\alias{GO2DecBP}
\title{GO biological process (BP) descendant GO-terms as an Edgelist}
\usage{
GO2DecBP(goterm)
}
\arguments{
\item{goterm}{A string object of BP GO-term}
}
\value{
A vector comprising of GO-terms and the nodes they are linked to
}
\description{
The GO2DecBP function provides a simple way to represent the biological process (BP) GO term.
             The function returns the descendant child nodes of a GO-term. In other words we begin from an
             ancestor term and find it's descendant child terms.
}
\examples{
v <- "GO:0000070" # BP GO term
GO2DecBP(v)
x <- "GO:0006725"
GO2DecBP(x)

}
