% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeListCc.R
\name{GO2DecCC}
\alias{GO2DecCC}
\title{GO cellular component (CC) descendant GO-terms as an Edgelist}
\usage{
GO2DecCC(goterm)
}
\arguments{
\item{goterm}{A string object of CC GO-terms}
}
\value{
A vector comprising of GO-terms and the nodes they are linked to
}
\description{
The GO2DecCC function provides a simple way to represent the cellular component (CC) GO-terms.
             The function returns the descendant child nodes of a GO-term. In other words we begin from an
             ancestor term and find it's descendant child terms.
}
\examples{
v <- "GO:0000799" # CC GO term
GO2DecCC(v)
x <- "GO:0043231"
GO2DecCC(x)

}
