% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeListMf.R
\name{GO2DecMF}
\alias{GO2DecMF}
\title{GO molecular function (MF) descendant GO-terms as an Edgelist}
\usage{
GO2DecMF(goterm)
}
\arguments{
\item{goterm}{A string object of MF GO-terms}
}
\value{
A vector comprising of GO-terms and the nodes they are linked to
}
\description{
The GO2DecMF function provides a simple way to represent the molecular function (MF) GO-terms.
             TThe function returns the descedant child nodes of a GO term.
             In other words we begin from an ancestor term and find it's descendant child terms.
}
\examples{
v <- "GO:0001228" # MF GO term
GO2DecMF(v)
x <- "GO:0005515"
GO2DecMF(x)

}
