% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeDistBP.R
\name{GOTermBP2ChildLevel}
\alias{GOTermBP2ChildLevel}
\title{Get the level of a biological process (BP) GO term's children}
\usage{
GOTermBP2ChildLevel(goterm)
}
\arguments{
\item{goterm}{A character string of a valid gene ontology id e.g "GO:0097278"}
}
\value{
A list of children terms and their respective levels
}
\description{
This function retrieves a GO BP term children's level i.e. for a GO-term it's children level are derived.
}
\examples{

# Retrieve "GO:0097278" children's level
GOTermBP2ChildLevel(goterm = "GO:0097278")

# Retrieve "GO:0051775" children's level
GOTermBP2ChildLevel(goterm = "GO:0051775")
}
