% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGoTree.R
\name{GOTermsApp.B}
\alias{GOTermsApp.B}
\title{Building tree one}
\usage{
GOTermsApp.B(t)
}
\arguments{
\item{t}{The GO root term}
}
\value{
returns GO terms
}
\description{
Building tree one
}
\keyword{internal}
