% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGoTree.R
\name{GOTermsApp.C}
\alias{GOTermsApp.C}
\title{Title}
\usage{
GOTermsApp.C(t)
}
\arguments{
\item{t}{The root GO term to build the GO tree}
}
\value{
Returns the GO tree for Cellular component
}
\description{
Title
}
\keyword{internal}
