% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGoTree.R
\name{GOTermsApp.M}
\alias{GOTermsApp.M}
\title{Title}
\usage{
GOTermsApp.M(t)
}
\arguments{
\item{t}{The root GO term to build the GO tree}
}
\value{
Returns the GO tree for Molecular function
}
\description{
Title
}
\keyword{internal}
