% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allterms.R
\name{maxLevelCC}
\alias{maxLevelCC}
\title{Return length of depth for cc}
\usage{
maxLevelCC(Organism)
}
\arguments{
\item{Organism}{The name of the organism}
}
\value{
length of depth
}
\description{
Return length of depth for cc
}
\keyword{internal}
