% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allterms.R
\name{maxLevelBP}
\alias{maxLevelBP}
\title{Return length of depth for bp}
\usage{
maxLevelBP(Organism)
}
\arguments{
\item{Organism}{The name of the organism}
}
\value{
length of depth
}
\description{
Return length of depth for bp
}
\keyword{internal}
