% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allterms.R
\name{maxLevelMF}
\alias{maxLevelMF}
\title{Return length of depth for mf}
\usage{
maxLevelMF(Organism)
}
\arguments{
\item{Organism}{The name of the organism}
}
\value{
length of depth
}
\description{
Return length of depth for mf
}
\keyword{internal}
